rename common.workmachine to common.work_machine
